# services/bank_account_web/src/models/account.ts
export class Account {
  account_number: number;
  balance: number;
  firstname: string;
  lastname: string;
  age: number;
  gender: string;
  address: string;
  employer: string;
  email: string;
  city: string;
  state: string
  constructor() {
  }

}